requirements/tune.txt
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
asttokens==2.4.0
attrs==23.1.0
ax-platform==0.3.4
backcall==0.2.0
botorch==0.9.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.0.4
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
distlib==0.3.7
exceptiongroup==1.1.3
executing==2.0.0
filelock==3.12.4
fonttools==4.43.1
frozenlist==1.4.0
gpytorch==1.11
grpcio==1.43.0
hydra-core==1.3.2
idna==3.4
imageio==2.31.5
imgaug==0.4.0
ipython==8.16.1
ipywidgets==8.1.1
jaxtyping==0.2.23
jedi==0.19.1
jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
lazy-loader==0.3
linear-operator==0.5.1
lmdb==1.4.1
markupsafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
msgpack==1.0.7
multipledispatch==1.0.0
networkx==3.1
numpy==1.26.1
omegaconf==2.3.0
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.0.1
platformdirs==3.11.0
plotly==5.17.0
prompt-toolkit==3.0.39
protobuf==3.20.3
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.16.1
pyparsing==3.1.1
pyro-api==0.1.2
pyro-ppl==1.8.6
python-dateutil==2.8.2
pytz==2023.3.post1
pyyaml==6.0.1
ray==1.13.0
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
shapely==2.0.2
six==1.16.0
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.2.3
tensorboardx==2.6.2.2
threadpoolctl==3.2.0
tifffile==2023.9.26
tqdm==4.66.1
traitlets==5.11.2
typeguard==2.13.3
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
virtualenv==20.24.5
wcwidth==0.2.8
widgetsnbextension==4.0.9
